+++ web/pgadmin/browser/server_groups/servers/databases/schemas/tables/templates/table/sql/10_plus/properties.sql
@@ -7,7 +7,7 @@ SELECT rel.oid, rel.relname AS name, rel.reltablespace AS spcoid,rel.relacl AS r
   (select nspname FROM pg_namespace WHERE oid = {{scid}}::oid ) as schema,
   pg_get_userbyid(rel.relowner) AS relowner, rel.relhasoids, rel.relkind,
   (CASE WHEN rel.relkind = 'p' THEN true ELSE false END) AS is_partitioned,
-  rel.relhassubclass, rel.reltuples, des.description, con.conname, con.conkey,
+  rel.relhassubclass, rel.reltuples::bigint, des.description, con.conname, con.conkey,
 	EXISTS(select 1 FROM pg_trigger
 			JOIN pg_proc pt ON pt.oid=tgfoid AND pt.proname='logtrigger'
 			JOIN pg_proc pc ON pc.pronamespace=pt.pronamespace AND pc.proname='slonyversion'
